+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 ## Changelog for Gmail Auto-forwarding
+- **Version 1.2 build 20240329**
+```
+- Add a new feature to validate the user's input
+- Add a trial option in the configuration
+```
 - **Version 1.1 build 20240317**
 ```
 - Fixed paging trigger sometimes not called